backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djoser==2.2.2
django-filter==25.1
sqlparse==0.5.3
tzdata==2025.2
pillow==11.2.1
reportlab==4.4.0
gunicorn==20.1.0 
psycopg2-binary==2.9.3
python-dotenv==1.1.0
django-extensions==3.2.3
